requirements.txt
asgiref
certifi
click
colorama
crispy-bootstrap5
distlib
Django
django-braces
django-crispy-forms
django-environ
django-filter
django-humanize
django-model-utils
django-star-ratings
djangorestframework
filelock
gunicorn
h11
humanize
Markdown
MouseInfo
numpy
opencv-contrib-python
packaging
pillow
pipenv
platformdirs
psycopg2
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
setuptools
sqlparse
swapper
tzdata
uvicorn
virtualenv